refactor: useCloseModal 훅으로 분리
