+++ src/hooks/index.ts
@@ -11,3 +11,4 @@ export * from './useRandomTitle';
 export * from './useGrid';
 export * from './useDebounce';
 export * from './useColorThief';
+export * from './useCloseModal';
