+++ src/pages/home/SearchPage.tsx
@@ -1,34 +1,16 @@
-import { gsap } from 'gsap';
-import { useCallback, useEffect, useRef, useState } from 'react';
 import ReactDOM from 'react-dom';
 
 import { MainClouds } from './MainClouds';
 import { SearchableBookList } from './SearchableBookList';
 
-import { useSetBackgroundColor } from 'hooks';
+import { useSetBackgroundColor, useCloseModal } from 'hooks';
 import { BookModal } from 'pages';
 import { useBookModalStore } from 'stores';
-import { handleModalClick } from 'utils';
 
 export const SearchPage = () => {
+  const { bookModalData } = useBookModalStore();
+  const { modalRef, closeModal } = useCloseModal();
   useSetBackgroundColor('bg-secondary', true);
-  const modalRef = useRef<HTMLDivElement>(null);
-  const { bookModalData, activeItem } = useBookModalStore();
-
-  const closeModal = useCallback(() => {
-    if (!modalRef.current) return;
-
-    let items: HTMLDivElement[];
-    if (activeItem.type === 'slider') {
-      items = gsap.utils.toArray<HTMLDivElement>('.wheel__card');
-    } else {
-      items = gsap.utils.toArray<HTMLDivElement>('.gallery__item');
-    }
-
-    if (items) {
-      handleModalClick(items, activeItem.index, modalRef.current);
-    }
-  }, [modalRef.current, activeItem.type, activeItem.index]);
 
   return (
     <div className='h-[calc(100vh-6rem)] w-full overflow-hidden relative'>
